Refactor: Move historymetadata classes in separate package
